First commit from repo migration
So all my previous commits are lost, unfortunately. I have been working on this site for a month or so now, but I made some mistakes with the large video files and uploading this to Github was just not working. There were large file storage issues, overages, upload limits, it was a full mess. Decided to start fresh with a clone of the repo without any video files in it at all, because they were lost inside the git history. So unfortunatelly they are no commits for this repo from the major development phase.

I can say that any code not written by me does have links referencing where I got the code ideas from (blogs, articles etc). I initially was using video.js but switched to reach player once I decided to just embed the videos rather than host them myself (terrible idea). Rowan's drag and drop page was the product of two weeks or so of prototyping to get to where it is now.

That's about it. If you're another developer reading this and have to work on this project, just message me if you need the context for a code decision. Happy to help!